Refactor ToTopButton
Add ToTopButton to Team Page

Minor Formatting

Change folder name to button
